+++ ModEntry.cs
@@ -20,19 +20,7 @@ namespace Gaphodil.BetterJukebox
          */
 
         /// <summary>The mod configuration from the player.</summary>
-        private ModConfig Config;
-
-        /// <summary>Internal music identifiers are displayed alongside the regular music name.</summary>
-        private bool ShowInternalId;
-
-        /// <summary>Ambience, sound effects, and other permanently disabled tracks show up in the jukebox.</summary>
-        private bool ShowAmbientTracks;
-
-        /// <summary>Songs not yet heard on the current save file can be found in the jukebox.</summary>
-        //private bool ShowUnheardTracks;
-
-        /// <summary>Non-default sorting options are enabled.</summary>
-        //private bool ShowAlternateSorts;
+        public ModConfig Config;
 
         /*
          * Public methods
@@ -47,13 +35,62 @@ namespace Gaphodil.BetterJukebox
             Config = helper.ReadConfig<ModConfig>();
 
             helper.Events.Display.MenuChanged += OnMenuChanged;
+            helper.Events.GameLoop.SaveLoaded += OnSaveLoaded;
             helper.Events.GameLoop.GameLaunched += OnGameLaunched;
         }
-
 
         /*
          * Private methods
          */
+
+        /// <summary>
+        /// Raised after loading a save.
+        /// </summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event data.</param>
+        private void OnSaveLoaded(object sender, SaveLoadedEventArgs e)
+        {
+            //Monitor.Log("all heard:\t" + string.Join(", ",Game1.player.songsHeard.ToArray()));
+            if (Config.PermanentUnheard)
+            {
+                // there was "don't repeat if already done"
+                // but it's gone now (because i added permanent blacklist)
+
+                if (Config.ShowUnheardTracks)
+                {
+                    BetterJukeboxHelper.AddUnheardTracks(
+                        Game1.player.songsHeard,
+                        Config.UnheardSoundtrack,
+                        Config.UnheardNamed,
+                        Config.UnheardRandom,
+                        Config.UnheardMisc,
+                        Config.UnheardDupes,
+                        Config.UnheardMusical
+                    );
+                }
+
+                //Game1.player.songsHeard.Remove("title_day"); // readded at every save load, so...
+                int MTIndex = Game1.player.songsHeard.IndexOf("MainTheme");
+                if (MTIndex.Equals(0) || MTIndex.Equals(-1)) { }
+                else
+                {
+                    Game1.player.songsHeard.RemoveAt(MTIndex);
+                    Game1.player.songsHeard.Insert(0, "MainTheme");
+                }
+
+                Monitor.Log("permanentBlacklist: " + Config.PermanentBlacklist);
+                Monitor.Log("permanentBlacklist converted: " + new FilterListConfig(Config.PermanentBlacklist));
+                FilterListConfig blacklist = new FilterListConfig(Config.PermanentBlacklist);
+                List<string> toRemove = blacklist.content.Distinct().ToList();
+                if (toRemove.Count > 0)
+                {
+                    foreach (string cue in toRemove)
+                    {
+                        Game1.player.songsHeard.Remove(cue);
+                    }
+                }
+            }
+        }
 
         /// <summary>
         /// Raised after a game menu is opened, closed, or replaced. 
@@ -62,14 +99,9 @@ namespace Gaphodil.BetterJukebox
         /// <param name="e">The event data.</param>
         private void OnMenuChanged(object sender, MenuChangedEventArgs e)
         {
-            // get config values here to reflect potential changes via GenericModConfigMenu
-            ShowInternalId = Config.ShowInternalId;
-            ShowAmbientTracks = Config.ShowAmbientTracks;
-            //ShowUnheardTracks = Config.ShowUnheardTracks;
-            //ShowAlternateSorts = Config.ShowAlternateSorts;
-
             // replace ChooseFromListMenu (only used for jukeboxes as of 1.4) with BetterJukeboxMenu
-            if (e.NewMenu is ChooseFromListMenu &&
+            if (Config.ShowMenu &&
+                e.NewMenu is ChooseFromListMenu &&
                 Helper.Reflection.GetField<bool>(e.NewMenu, "isJukebox").GetValue() == true)
             {
                 ChooseFromListMenu.actionOnChoosingListOption action = 
@@ -78,36 +110,49 @@ namespace Gaphodil.BetterJukebox
                 e.NewMenu.exitThisMenuNoSound(); // is this neccessary? is there a better way?
 
                 // create default list of songs to play - apparently this is how CA hard-copied the list
-                List<string> list = Game1.player.songsHeard.Distinct<string>().ToList<string>();
+                List<string> list = Game1.player.songsHeard.Distinct().ToList();
 
-                //if (ShowUnheardTracks)
-                //{
-                //    AddAllTracks(list);
-                //}
-
-                // from ChooseFromListMenu constructor
-                if (!ShowAmbientTracks)
+                // add unheard tracks
+                if (Config.ShowUnheardTracks && !Config.PermanentUnheard)
                 {
-                    RemoveAmbience(list);
+                    BetterJukeboxHelper.AddUnheardTracks(
+                        list,
+                        Config.UnheardSoundtrack,
+                        Config.UnheardNamed,
+                        Config.UnheardRandom,
+                        Config.UnheardMisc,
+                        Config.UnheardDupes,
+                        Config.UnheardMusical
+                    );
                 }
-                //else  // rain noises heavily based on actual weather and player location, no effect if not raining
-                //{
-                //    if (list.IndexOf("rain") == -1)
-                //        list.Add("rain");
-                //}
+
+                // remove specific tracks
+                BetterJukeboxHelper.FilterTracksFromList(list, Config.AmbientTracks, Config.Blacklist, Config.Whitelist);
+
                 list.Remove("title_day"); // this one gets removed for A Good Reason, apparently
 
                 // this is the one change that isn't true to how the game does it, because it makes me angy >:L
                 int MTIndex = list.IndexOf("MainTheme");
-                if (MTIndex.Equals(0)) { }
-                else {
+                // if permanent: not -1
+                // if not permanent: -1
+                // if in either blacklist: -1
+                // impossible to tell directly, can't put into filtertracks because need to check both
+                bool isMTPermanentBlacklisted = new FilterListConfig(Config.PermanentBlacklist).content.Contains("MainTheme");
+                bool isMTRegularBlacklisted = new FilterListConfig(Config.Blacklist).content.Contains("MainTheme");
+
+                if (!isMTPermanentBlacklisted &&
+                    !isMTRegularBlacklisted &&
+                    !MTIndex.Equals(0))
+                {
                     if (MTIndex.Equals(-1)) { }
-                    else list.RemoveAt(MTIndex);
+                    else
+                        list.RemoveAt(MTIndex);
                     list.Insert(0, "MainTheme"); 
                 }
 
                 // speculative fix for Nexus page bug report
                 list.Remove("resetVariable");
+
 
                 // create and activate the menu
                 Game1.activeClickableMenu = new BetterJukeboxMenu(
@@ -119,8 +164,8 @@ namespace Gaphodil.BetterJukebox
                     ),
                     key => Helper.Translation.Get(key),
                     Monitor,
-                    Game1.player.currentLocation.miniJukeboxTrack.Value,
-                    ShowInternalId
+                    Config,
+                    Game1.player.currentLocation.miniJukeboxTrack.Value
                 ); 
             }
         }
@@ -146,74 +191,158 @@ namespace Gaphodil.BetterJukebox
             // add some config options
             api.RegisterSimpleOption(
                 mod: ModManifest,
-                optionName: "Show Internal ID",
-                optionDesc: "Internal music identifiers are displayed alongside the regular music name.",
+                optionName: Helper.Translation.Get("BetterJukebox:ShowMenu"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowMenuDescription"),
+                optionGet: () => Config.ShowMenu,
+                optionSet: value => Config.ShowMenu = value
+            );
+            api.RegisterPageLabel(
+                ModManifest,
+                Helper.Translation.Get("BetterJukebox:ListSettings"),
+                Helper.Translation.Get("BetterJukebox:ListSettingsDescription"),
+                Helper.Translation.Get("BetterJukebox:ListSettings")
+            );
+            api.RegisterPageLabel(
+                ModManifest,
+                Helper.Translation.Get("BetterJukebox:FunctionalSettings"),
+                Helper.Translation.Get("BetterJukebox:FunctionalSettingsDescription"),
+                Helper.Translation.Get("BetterJukebox:FunctionalSettings")
+            );
+            api.RegisterPageLabel(
+                ModManifest,
+                Helper.Translation.Get("BetterJukebox:VisualSettings"),
+                Helper.Translation.Get("BetterJukebox:VisualSettingsDescription"),
+                Helper.Translation.Get("BetterJukebox:VisualSettings")
+            );
+
+            api.StartNewPage(ModManifest, Helper.Translation.Get("BetterJukebox:ListSettings"));
+            api.RegisterClampedOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:AmbientTracks"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:AmbientTracksDescription"),
+                optionGet: () => Config.AmbientTracks,
+                optionSet: value => Config.AmbientTracks = value,
+                min: 0,
+                max: 2,
+                interval: 1
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:Blacklist"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:BlacklistDescription"),
+                optionGet: () => Config.Blacklist,
+                optionSet: value => Config.Blacklist = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:Whitelist"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:WhitelistDescription"),
+                optionGet: () => Config.Whitelist,
+                optionSet: value => Config.Whitelist = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:ShowLockedSongs"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowLockedSongsDescription"),
+                optionGet: () => Config.ShowLockedSongs,
+                optionSet: value => Config.ShowLockedSongs = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:ShowUnheardTracks"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowUnheardTracksDescription"),
+                optionGet: () => Config.ShowUnheardTracks,
+                optionSet: value => Config.ShowUnheardTracks = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardSoundtrack"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardSoundtrackDescription"),
+                optionGet: () => Config.UnheardSoundtrack,
+                optionSet: value => Config.UnheardSoundtrack = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardNamed"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardNamedDescription"),
+                optionGet: () => Config.UnheardNamed,
+                optionSet: value => Config.UnheardNamed = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardRandom"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardRandomDescription"),
+                optionGet: () => Config.UnheardRandom,
+                optionSet: value => Config.UnheardRandom = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardMisc"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardMiscDescription"),
+                optionGet: () => Config.UnheardMisc,
+                optionSet: value => Config.UnheardMisc = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardDupes"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardDupesDescription"),
+                optionGet: () => Config.UnheardDupes,
+                optionSet: value => Config.UnheardDupes = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:UnheardMusical"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:UnheardMusicalDescription"),
+                optionGet: () => Config.UnheardMusical,
+                optionSet: value => Config.UnheardMusical = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:PermanentUnheard"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:PermanentUnheardDescription"),
+                optionGet: () => Config.PermanentUnheard,
+                optionSet: value => Config.PermanentUnheard = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:PermanentBlacklist"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:PermanentBlacklistDescription"),
+                optionGet: () => Config.PermanentBlacklist,
+                optionSet: value => Config.PermanentBlacklist = value
+            );
+
+            api.StartNewPage(ModManifest, Helper.Translation.Get("BetterJukebox:FunctionalSettings"));
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:TrueRandom"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:TrueRandomDescription"),
+                optionGet: () => Config.TrueRandom,
+                optionSet: value => Config.TrueRandom = value
+            );
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:ShowAlternateSorts"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowAlternateSortsDescription"),
+                optionGet: () => Config.ShowAlternateSorts,
+                optionSet: value => Config.ShowAlternateSorts = value
+            );
+
+            api.StartNewPage(ModManifest, Helper.Translation.Get("BetterJukebox:VisualSettings"));
+            api.RegisterSimpleOption(
+                mod: ModManifest,
+                optionName: Helper.Translation.Get("BetterJukebox:ShowInternalId"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowInternalIdDescription"),
                 optionGet: () => Config.ShowInternalId,
                 optionSet: value => Config.ShowInternalId = value
             );
             api.RegisterSimpleOption(
                 mod: ModManifest,
-                optionName: "Show Ambient Tracks",
-                optionDesc: "Ambience, sound effects, and other permanently disabled tracks show up in the jukebox.",
-                optionGet: () => Config.ShowAmbientTracks,
-                optionSet: value => Config.ShowAmbientTracks = value
+                optionName: Helper.Translation.Get("BetterJukebox:ShowBandcampNames"),
+                optionDesc: Helper.Translation.Get("BetterJukebox:ShowBandcampNamesDescription"),
+                optionGet: () => Config.ShowBandcampNames,
+                optionSet: value => Config.ShowBandcampNames = value
             );
-            //api.RegisterSimpleOption(
-            //    mod: this.ModManifest,
-            //    optionName: "Show Unheard Tracks",
-            //    optionDesc: "Songs not yet heard on the current save file can be found in the jukebox.",
-            //    optionGet: () => this.Config.ShowUnheardTracks,
-            //    optionSet: value => this.Config.ShowUnheardTracks = value
-            //);
-            //api.RegisterSimpleOption(
-            //    mod: this.ModManifest,
-            //    optionName: "Show Alternate Sorts",
-            //    optionDesc: "Non-default sorting options are enabled.",
-            //    optionGet: () => this.Config.ShowAlternateSorts,
-            //    optionSet: value => this.Config.ShowAlternateSorts = value
-            //);
         }
 
-        /// <summary>
-        /// Remove ambient tracks from the list of songs available in the jukebox.
-        /// Copied from the ChooseFromListMenu constructor.
-        /// </summary>
-        /// <param name="trackList"></param>
-        private void RemoveAmbience(List<string> trackList)
-        {
-            for (int index = trackList.Count - 1; index >= 0; --index)
-            {
-                if (trackList[index].ToLower().Contains("ambient") || trackList[index].ToLower().Contains("bigdrums") || trackList[index].ToLower().Contains("clubloop") || trackList[index].ToLower().Contains("ambience"))
-                {
-                    trackList.RemoveAt(index);
-                }
-                else
-                {
-                    switch (trackList[index])
-                    {
-                        case "buglevelloop": // vanilla bug: should be "bugLevelLoop"
-                            trackList.RemoveAt(index);
-                            continue;
-                        case "coin":
-                            trackList.RemoveAt(index);
-                            continue;
-                        case "communityCenter":
-                            trackList.RemoveAt(index);
-                            continue;
-                        case "jojaOfficeSoundscape":
-                            trackList.RemoveAt(index);
-                            continue;
-                        case "nightTime":
-                            trackList.RemoveAt(index);
-                            continue;
-                        case "ocean":
-                            trackList.RemoveAt(index);
-                            continue;
-                        default:
-                            continue;
-                    }
-                }
-            }
-        }
     }
 }